scolv/ focal sphere/ quadrants colours : darker?
As discussed in https://forum.seiscomp.de/t/scv4-scolv-focal-sphere-quadrants-colours , the colour of the unconfirmed focal sphere compression quadrants is a bit confusing. Keeping the same kind of colour than original, there are 3 other possibilities :  « Dark », « Mid » and « Midlight ». I propose to use « Mid ». 

I understand the confirmed colour is « Highlight » with a 25% blend with background color, and won't be changed.