[trunk] Support double precision records for display
Double precision samples were converted to float for display purposes
and could be shown in many cases. But if a double precision value does
not fit into a float it would lead to incorrect plots. This commit
allows to render double precision records with full precision.

Some Qt signals and slots were required to change its signature from
float to double which is not checked at compile time. This would have
an effect on external code which uses the affected signals/slots. See
version.h for more information.