[system] Filter invalid unknown commandline options
This affects how an application can be run. Every unknown
command line option starting with '-' will be checked.
Unknown means that it is not part of the description (--help).
If the unknown argument does not start with "--" or does not
contain an equal sign "=" then an error will be thrown that
this option is not known.