[datamodel] Fix PublicObjectCache::find method and add PublicObjectCache::typeInfo
The find method just returned a cached object if it is part of the
global registry. This makes sense in most of the cases but does not work
if registration is disabled and objects are just cached without the
global pool. This commit adds the lookup into the cache storage if the
registry does not hold the requested object.