[xml] Improve performance in scml to quakeml XSLT
SCML documents describing large earthquakes were taking a long time to
convert to QuakeML, with almost all that time being spent in the loop
that outputs picks that aren't referenced by amplitudes. (For example,
an event with ~600 arrivals in its preferred origin was taking 12
seconds.)

By precomputing the list of referenced amplitudes once per origin
(rather than computing it on the fly for each pick), performance is
greatly improved (down to 400ms in my 600-arrival example).