[stdloc] better solutions for not well constrained events
This is a minor change for well constrained locations but  it has a huge
impact on events with small amount of picks or with outliers.  That is
important when using stdloc in combination with scanloc for example.

Also lower level of log messages, since they might be useful only when
debugging.