[stdloc] handle missing ttt for certain phases
The previous behaviour was to throw and exception and avoid
locating an event when it was not possible to fetch the ttt
for a phase. With this patch, the phases (arrivals) for which
the ttt is not availale are simply disabled and the location
is performed on the remaining phases.