[stdloc] OctTree search optimization
Performed estensive comparison tests with NonLinLoc and the current
code gave results in line with NLL using similar settings (max num
iterations, min cell size, gaussian L2 misfit). The previous code
was slower to converge to the same solutions, but the solutions
are consistent with the new ones if max iterations is increased.