[stdloc] set enableConfidenceEllipsoid to false by default
The common usage of stdloc is to be used by scanloc and in this case
the confidence ellipsoid is not required.

More importantly, I do not trust the eigv/chi2 code: if that code crashes
when called from scolv that is annoying, but the user notices the issue.
However, If the code crashes when called by scanloc than who is going to
notice that?