[datamodel] Update to 0.13.2
This update changes database column types from blob to text.
Blobs (in particular PostgreSQL bytea) requires a special syntax
for encoding normal text. As we do not save blobs but normal text
this is semantically wrong anyway. To avoid encoding errors and
to fix the wrong semantics, this commit fixes this particular issue.

Outstanding issues are the remarks in many types which are still
of datamodel type "Blob" which needs to be corrected in the future
as well.

We skip to data model version 0.13.2 because internally we already
had 0.13.1 as intermediate version which still needs to be
identifiable.