[swig] Add NumPy include first
Until now, the CMake variable PYTHON_NUMPY_INCLUDE_DIR had no effect
on systems where NumPy was installed in the Python include dir and the
variable was set to a different directory. In this case the build
always used the NumPy installation from the include dir
because it had a higher priority. This commit ensures that the variable
PYTHON_NUMPY_INCLUDE_DIR is taken into account.