[gui] Add projection auto-detection
As the projection of the tiles as well as the projection of the
rendered map has to be configured to use Mercator map projection,
this commit adds auto-detection which uses the projection hint
of the TileStore. The default implementation of the TileStore
directory implements detection based on image width/height ratio.